chore: move defaultSections objects into NewSectionCreator
